+++ src/components/icon-button/icon-button.tsx
@@ -3,9 +3,10 @@ import { Component, Element, h, Method, Prop } from '@stencil/core';
 import { IconSize } from '@limetech/lime-elements';
 
 /**
- * @exampleComponent limel-example-icon-button
- * @exampleComponent limel-example-icon-button-toggle-state
+ * @exampleComponent limel-example-icon-button-basic
  * @exampleComponent limel-example-icon-button-elevated
+ * @exampleComponent limel-example-icon-button-toggle-state
+ * @exampleComponent limel-example-icon-button-composite
  */
 @Component({
     tag: 'limel-icon-button',